add : success get form value for passengers field, clean code, fixed position breadcrumb timer, total price, loading skeleton init
