+++ src/components/Booking/index.jsx
@@ -2,5 +2,14 @@ import BreadcrumbWithTimer from "./BreadcrumbWithTimer";
 import FlightDetail from "./FlightDetail";
 import FormSchema from "./FormSchema";
 import SeatPicker from "./SeatPicker";
+import OrdererField from "./OrdererField";
+import BookingForm from "./BookingForm";
 
-export { BreadcrumbWithTimer, FlightDetail, FormSchema, SeatPicker };
+export {
+  BreadcrumbWithTimer,
+  FlightDetail,
+  FormSchema,
+  SeatPicker,
+  OrdererField,
+  BookingForm,
+};
