+++ src/components/ui/skeleton.jsx
@@ -1,15 +1,14 @@
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib/utils"
 
-function Skeleton({ className, ...props }) {
+function Skeleton({
+  className,
+  ...props
+}) {
   return (
-    <div
-      className={cn(
-        "animate-pulse rounded-md bg-slate-200 dark:bg-slate-800",
-        className
-      )}
-      {...props}
-    />
+    (<div
+      className={cn("animate-pulse rounded-md bg-slate-100 dark:bg-slate-800", className)}
+      {...props} />)
   );
 }
 
-export { Skeleton };
+export { Skeleton }
